Split workflow_template into submodules to isolate jobmon dependency
